example/agentic-recommendation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
motor==3.3.2
pydantic==2.5.3
pymongo==4.6.1
python-dotenv==1.0.0
aiokafka==0.10.0
pandas==2.1.4
python-multipart==0.0.6
sqlmodel==0.0.14
aiosqlite==0.19.0
redis==5.0.1
aioredis==2.0.1
openai>=1.12.0
httpx>=0.27.0
aiofiles==23.2.1
python-dateutil==2.8.2

